I moved the interpolator classes for flatfield events to to their proper place
