+++ src/ctapipe/monitoring/interpolation.py
@@ -11,6 +11,77 @@ from ctapipe.core import Component, traits
 
 from .astropy_helpers import read_table
 
+class ChunkFunction:
+
+    """
+    Chunk Interpolator for the gain and pedestals
+    Interpolates data so that for each time the value from the latest starting
+    valid chunk is given or the earliest available still valid chunk for any
+    pixels without valid data.
+
+    Parameters
+    ----------
+    values : None | np.array
+        Numpy array of the data that is to be interpolated.
+        The first dimension needs to be an index over time
+    times : None | np.array
+        Time values over which data are to be interpolated
+        need to be sorted and have same length as first dimension of values
+    """
+
+    def __init__(
+        self,
+        start_times,
+        end_times,
+        values,
+        bounds_error=True,
+        fill_value="extrapolate",
+        assume_sorted=True,
+        copy=False,
+    ):
+        self.values = values
+        self.start_times = start_times
+        self.end_times = end_times
+        self.bounds_error = bounds_error
+        self.fill_value = fill_value
+
+    def __call__(self, point):
+        if point < self.start_times[0]:
+            if self.bounds_error:
+                raise ValueError("below the interpolation range")
+
+            if self.fill_value == "extrapolate":
+                return self.values[0]
+
+            else:
+                a = np.empty(self.values[0].shape)
+                a[:] = np.nan
+                return a
+
+        elif point > self.end_times[-1]:
+            if self.bounds_error:
+                raise ValueError("above the interpolation range")
+
+            if self.fill_value == "extrapolate":
+                return self.values[-1]
+
+            else:
+                a = np.empty(self.values[0].shape)
+                a[:] = np.nan
+                return a
+
+        else:
+            i = np.searchsorted(
+                self.start_times, point, side="left"
+            )  # Latest valid chunk
+            j = np.searchsorted(
+                self.end_times, point, side="left"
+            )  # Earliest valid chunk
+            return np.where(
+                np.isnan(self.values[i - 1]), self.values[j], self.values[i - 1]
+            )  # Give value for latest chunk unless its nan. If nan give earliest chunk value
+
+
 class Interpolator(Component, metaclass=ABCMeta):
     """
     Interpolator parent class.
@@ -182,3 +253,119 @@ class PointingInterpolator(Interpolator):
         self._interpolators[tel_id] = {}
         self._interpolators[tel_id]["az"] = interp1d(mjd, az, **self.interp_options)
         self._interpolators[tel_id]["alt"] = interp1d(mjd, alt, **self.interp_options)
+
+
+class FlatFieldInterpolator(Interpolator):
+    """
+    Interpolator for flatfield data
+    """
+
+    telescope_data_group = "dl1/calibration/gain"  # TBD
+    required_columns = frozenset(["start_time", "end_time", "gain"])
+    expected_units = {"gain": u.one}
+
+    def __call__(self, tel_id, time):
+        """
+        Interpolate flatfield data for a given time and tel_id.
+
+        Parameters
+        ----------
+        tel_id : int
+            telescope id
+        time : astropy.time.Time
+            time for which to interpolate the calibration data
+
+        Returns
+        -------
+        ffield : array [float]
+            interpolated flatfield data
+        """
+
+        self._check_interpolators(tel_id)
+
+        ffield = self._interpolators[tel_id]["gain"](time)
+        return ffield
+
+    def add_table(self, tel_id, input_table):
+        """
+        Add a table to this interpolator
+
+        Parameters
+        ----------
+        tel_id : int
+            Telescope id
+        input_table : astropy.table.Table
+            Table of pointing values, expected columns
+            are ``time`` as ``Time`` column and "gain"
+            for the flatfield data
+        """
+
+        self._check_tables(input_table)
+
+        input_table = input_table.copy()
+        input_table.sort("start_time")
+        start_time = input_table["start_time"]
+        end_time = input_table["end_time"]
+        gain = input_table["gain"]
+        self._interpolators[tel_id] = {}
+        self._interpolators[tel_id]["gain"] = ChunkFunction(
+            start_time, end_time, gain, **self.interp_options
+        )
+
+
+class PedestalInterpolator(Interpolator):
+    """
+    Interpolator for Pedestal data
+    """
+
+    telescope_data_group = "dl1/calibration/pedestal"  # TBD
+    required_columns = frozenset(["start_time", "end_time", "pedestal"])
+    expected_units = {"pedestal": u.one}
+
+    def __call__(self, tel_id, time):
+        """
+        Interpolate pedestal or gain for a given time and tel_id.
+
+        Parameters
+        ----------
+        tel_id : int
+            telescope id
+        time : astropy.time.Time
+            time for which to interpolate the calibration data
+
+        Returns
+        -------
+        pedestal : array [float]
+            interpolated pedestal values
+        """
+
+        self._check_interpolators(tel_id)
+
+        pedestal = self._interpolators[tel_id]["pedestal"](time)
+        return pedestal
+
+    def add_table(self, tel_id, input_table):
+        """
+        Add a table to this interpolator
+
+        Parameters
+        ----------
+        tel_id : int
+            Telescope id
+        input_table : astropy.table.Table
+            Table of pointing values, expected columns
+            are ``time`` as ``Time`` column and "pedestal"
+            for the pedestal data
+        """
+
+        self._check_tables(input_table)
+
+        input_table = input_table.copy()
+        input_table.sort("start_time")
+        start_time = input_table["start_time"]
+        end_time = input_table["end_time"]
+        pedestal = input_table["pedestal"]
+        self._interpolators[tel_id] = {}
+        self._interpolators[tel_id]["pedestal"] = ChunkFunction(
+            start_time, end_time, pedestal, **self.interp_options
+        )
